requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
# SDK oficial de Brevo Python (única librería de email)
brevo-python
